change move swagger command code out of cli file and into individual modual
